Version - Version 1.0. Started working on the Header component.
js/jsx - In the component App.jsx, the styles file and component Header.jsx were imported. In App.jsx, the main tags of the website were spelled out. The Header.jsx component was created, the main tags were added to it and the styles module was connected.

css/scss - In the file App.scss, were added to the main css properties and grid elements to work with the project. In Header.jsx, were added to the main css properties and flex elements.

img - Images have been added, to work with content.

Translated with www.DeepL.com/Translator (free version)