Version 1.1. Creating the Experience.jsx component.
js/jsx - In App.jsx, the Experience.jsx component was imported. The Experience.jsx component was created and basic tags and css modules were added to it.

css/scss - In the file Experience.scss, were added the main classes to work with the component.

img - Images were added to work with component.