Version 1.2. Work has begun on the Feedback.jsx component.
js/jsx - In the main component of App.js, was imported and added to the code structure, component Feedback.jsx. The Feedback.jsx component and the files to work with this component were created. A basic page structure was written, a library for animations and a styles file were imported.

css/scss - A styles file was created. With the help of importan, the basic styles of the website were added.

img - Images were added, to work with the component.