Version 1.3. Work has begun on the Allien.jsx component.
js/jsx - The file Allien.jsx was created. A folder for images and style files was created to it. In the component were added variables, which contains the animation. The "Framer-Motion" library was imported. The main skeleton of the page was written.

css/scss - Basic site styles were imported.

img - Images were added, to work with the component.