The version is 1.3.1. Work on the component Allien.jsx.
js/jsx - The main content for the component was added. Basic tags were added and BEM classes were added to them. Images, a styles file and the Framer-Motion library were imported. The use of css modules was eliminated.

css/scss - Normal css files have been added to all components. Refusal of further work with CSS modules in favor of BEM, because of greater convenience and speed of writing code. For the component Allien.jsx, the basic page styles were set.