Version - 1.4. Starting work on the Prices.jsx component.
js/jsx - The file Prices.jsx was created. In it the basic libraries for work were added, the styles file was imported and images were imported.

css/scss - In the styles files, the basic classes for working with the component were added. The main parameters of this block were set.

img - Images were added to work with the component.