Version 1.4.2. Work on the Prices.jsx component.
js/jsx - All the work on the HTML structure of the page was completed. All the basic tags and content were added to the component. During the work, the code was slightly redone and additional classes were added.

css/scss - The styles file was completely made. All the necessary CSS style properties were added and pseudoclasses were used to create animations when hovering over the content. Flex and grid elements were used.

img - Images were added to work with the component.