Version 1.5. Starting work on the About.jsx component.
js/jsx - The About.jsx component was created. App.jsx was added to the main block of the site and was imported into this component. In About.jsx the main blocks for working with tags and content were added.

css/scss - Styles files were created and basic site styles were imported.

img - Images were added, to work with the block.