Version 1.6. Starting work on the Footer.jsx component.
js/jsx - The Footer.jsx component was created, it was imported and added to the code, the main App.jsx component. In the Footer.jsx component, basic blocks and classes were created to work with tags. A library with animations was added, a styles file was connected and images were imported.

css/scss - Styles files were created. The main parameters of the site were imported into them.

img - Images were added, to work with the block.